deprecated SuspendableAsset
