+++ android/player/src/main/java/com/intuit/playerui/android/compose/ComposableAsset.kt
@@ -18,14 +18,16 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.viewinterop.AndroidView
 import com.intuit.playerui.android.AssetContext
+import com.intuit.playerui.android.asset.DecodableAsset
 import com.intuit.playerui.android.asset.RenderableAsset
-import com.intuit.playerui.android.asset.SuspendableAsset
 import com.intuit.playerui.android.build
 import com.intuit.playerui.android.extensions.Styles
 import com.intuit.playerui.android.extensions.into
 import com.intuit.playerui.android.withContext
 import com.intuit.playerui.android.withTag
 import com.intuit.playerui.core.experimental.ExperimentalPlayerApi
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import kotlinx.serialization.KSerializer
 
@@ -39,17 +41,13 @@ import kotlinx.serialization.KSerializer
 public abstract class ComposableAsset<Data> (
     assetContext: AssetContext,
     serializer: KSerializer<Data>,
-) : SuspendableAsset<Data>(assetContext, serializer) {
+) : DecodableAsset<Data>(assetContext, serializer) {
 
     override suspend fun initView(data: Data) = ComposeView(requireContext()).apply {
         layoutParams = if (asset is ViewportAsset) {
             ViewGroup.LayoutParams(MATCH_PARENT, MATCH_PARENT)
         } else {
             ViewGroup.LayoutParams(WRAP_CONTENT, WRAP_CONTENT)
-        }
-
-        setContent {
-            compose(data = data)
         }
     }
 
@@ -110,6 +108,8 @@ private fun RenderableAsset.composeAndroidView(
     styles: Styles? = null,
 ) {
     AndroidView(factory = ::FrameLayout, modifier) {
-        render(styles) into it
+        this.hydrationScope.launch {
+            render(styles) into it
+        }
     }
 }
