+++ android/player/src/main/java/com/intuit/playerui/android/ui/PlayerFragment.kt
@@ -15,8 +15,8 @@ import androidx.lifecycle.repeatOnLifecycle
 import androidx.lifecycle.whenStarted
 import androidx.transition.Transition
 import com.intuit.playerui.android.AndroidPlayer
+import com.intuit.playerui.android.asset.DecodableAsset
 import com.intuit.playerui.android.asset.RenderableAsset
-import com.intuit.playerui.android.asset.SuspendableAsset
 import com.intuit.playerui.android.extensions.into
 import com.intuit.playerui.android.extensions.transitionInto
 import com.intuit.playerui.android.lifecycle.ManagedPlayerState
@@ -153,7 +153,7 @@ public abstract class PlayerFragment : Fragment(), ManagedPlayerState.Listener {
         val startTime = System.currentTimeMillis()
         val view = asset?.render(requireContext())?.let {
             // unwrap if we know we have an async view stub, and just wait on the actual view
-            if (it is SuspendableAsset.AsyncViewStub) it.awaitView() else it
+            if (it is DecodableAsset.AsyncViewStub) it.awaitView() else it
         }
 
         view?.doOnLayout {
@@ -182,7 +182,7 @@ public abstract class PlayerFragment : Fragment(), ManagedPlayerState.Listener {
         renderingJob?.cancel("handling new update")
         renderingJob = lifecycleScope.launch {
             whenStarted { // TODO: This'll go away when we can call a suspend version of this
-                withContext(if (asset is SuspendableAsset<*>) Dispatchers.Default else Dispatchers.Main) {
+                withContext(Dispatchers.Default) {
                     try {
                         renderIntoPlayerCanvas(asset, animateTransition)
                     } catch (exception: Exception) {
