+++ android/player/src/test/java/com/intuit/playerui/android/utils/OtherSimpleAsset.kt
@@ -3,7 +3,7 @@ package com.intuit.playerui.android.utils
 import android.view.View
 import android.widget.ImageView
 import com.intuit.playerui.android.AssetContext
-import com.intuit.playerui.android.asset.SuspendableAsset
+import com.intuit.playerui.android.asset.DecodableAsset
 import com.intuit.playerui.core.asset.Asset
 import com.intuit.playerui.core.bridge.Node
 import com.intuit.playerui.core.bridge.runtime.serialize
@@ -14,7 +14,7 @@ import com.intuit.playerui.utils.makeFlow
 import kotlinx.serialization.json.Json
 
 @Suppress("DEPRECATION_ERROR")
-internal class OtherSimpleAsset(assetContext: AssetContext) : SuspendableAsset<Node>(assetContext, NodeSerializer()) {
+internal class OtherSimpleAsset(assetContext: AssetContext) : DecodableAsset<Node>(assetContext, NodeSerializer()) {
 
     override suspend fun initView(data: Node) = ImageView(context)
