+++ plugins/reference-assets/android/src/main/java/com/intuit/playerui/android/reference/assets/input/Input.kt
@@ -4,8 +4,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.inputmethod.EditorInfo
 import com.intuit.playerui.android.AssetContext
+import com.intuit.playerui.android.asset.DecodableAsset
 import com.intuit.playerui.android.asset.RenderableAsset
-import com.intuit.playerui.android.asset.SuspendableAsset
 import com.intuit.playerui.android.extensions.into
 import com.intuit.playerui.android.reference.assets.R
 import com.intuit.playerui.android.reference.assets.text.Text
@@ -17,7 +17,7 @@ import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withContext
 import kotlinx.serialization.Serializable
 
-class Input(assetContext: AssetContext) : SuspendableAsset<Input.Data>(assetContext, Data.serializer()) {
+class Input(assetContext: AssetContext) : DecodableAsset<Input.Data>(assetContext, Data.serializer()) {
 
     @Serializable
     data class Validation(
